Move search bar to core
